Refactor: collector now have own folder
