+++ collector/internal/domain/repository.go
